feat: move router to separate folder
